Moved LinearFitting to base objective file
