+++ cofi/cofi_objective/__init__.py
@@ -1,11 +1,12 @@
 from .model_params import Model
-from .base_objective import BaseObjective, LeastSquareObjective
+from .base_objective import BaseObjective, LeastSquareObjective, LinearFittingObjective
 from .base_forward import BaseForward, LinearFittingFwd, PolynomialFittingFwd
 
 __all__ = [
     "Model",
     "BaseObjective",
     "LeastSquareObjective",
+    "LinearFittingObjective",
     "BaseForward",
     "LinearFittingFwd",
     "PolynomialFittingFwd",
