+++ cofi/cofi_objective/examples/__init__.py
@@ -1,5 +1,4 @@
 from .exponential_decay import ExpDecay
 from .bandit_multi_armed import MultiArmedBandit
-from .linear_fitting import LinearFitting
 
-__all__ = ["ExpDecay", "MultiArmedBandit", "LinearFitting"]
+__all__ = ["ExpDecay", "MultiArmedBandit"]
